Criação do StoreUserRequest para escrever as regras de validação e mensagens personalizadas do CRUD de usuarios.

+++ app/Http/Controllers/UserController.php
@@ -4,9 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\MessageBag;
-use App\Http\Requests\StoreUser;
+use App\Http\Requests\StoreUserRequest;
+
 
 class UserController extends Controller
 {
@@ -15,49 +14,62 @@ class UserController extends Controller
     {
         $all_users = User::all();
 
-        return response()->json(['all_users' => $all_users], 200);
+        return response()->json(['user' => $all_users], 200);
     }
 
 
     // Função para criar usuarios.
-    public function createUser(StoreUser $request) 
+    public function createUser(StoreUserRequest $request) // 
     {
-        $validated = $request->validated();
-
-        // Mostrando os erros que foram gerados
-        /*if ($validated->fails()) {
-            $errors = $validatedData->errors();
-            
-            return response()->json(['errors' => $errors], 500); 
-        }*/
-
         $new_user = User::create($request->all());
-        return response()->json(['created_user' => $new_user], 201);
 
+        return response()->json(['user' => $new_user], 201);
     }
 
 
-    // Buscar usuario por ID
-    public function getUser(User $id) 
+    // Buscar usuario por ID - 1
+    public function getUser($id) 
     {
-        return response()->json(['id_user' => $id], 200);
+        $get_user = User::find($id);
+        
+        if (is_null($get_user)) {
+            return response()->json(['message' => 'Usuario nao encontrado.'], 404);
+        }
+
+        return response()->json(['user' => $get_user], 200);
     }
 
+    /* Buscar usuario por ID - 2 || Essa função é um modo mais enxuto de fazer a requisição. Optei pela função de cima para manter um padrão de retorno no código.
+    public function get_User(User $id)
+    {
+        return response()->json(['user' => $id], 200);
+    }*/
 
-    // Editar atributos do usuario
-    public function updateUser(Request $request, $id)
-    {   
-        $updated_user = User::findOrFail($id); // Buscando o ID cadastrado
-        $updated_user->update($request->all());
+    // Funçao para editar atributos do usuario
+    public function updateUser(StoreUserRequest $request, $id)
+    {    
+        $updated_user = User::find($id); // Buscando o ID cadastrado
         
-        return response()->json(['updated_user'=> $updated_user], 200);
+        if(is_null($updated_user)){
+            return response()->json(['message' => 'Usuario nao encontrado.'], 404); // Verificando se o usuario existe.
+        }
+        
+        $validated = $request->validated(); // Validando informaçoes
+        $updated_user->update($request->all()); // Atualizando dados
+    
+        return response()->json(['user' => $updated_user], 200);
     } 
         
     
-    // Deletar usuario
+    // Função para deletar usuario
     public function deleteUser($id)
     {
-        $user = User::findOrFail($id);
+        $user = User::find($id);
+        
+        if (is_null($user)) {
+            return response()->json(['message' => 'Usuario nao encontrado.'], 404);
+        }
+
         $user->delete();
 
         return response()->json(['message'=> 'Usuario deletado com sucesso.'], 200);
